+++ Components/Core/Conditions/ConditionObject.lua
@@ -426,40 +426,79 @@ function ConditionObject:IsUnitAuraEventRelevant(updatedAuras, spell, onlyMine)
 	end
 end
 
+function ConditionObject:IsTmwAuraEventRelevant(updatedAuras, spellSet, onlyMine)
+	-- updatedAuras = { [name | id | dispelType] = mightBeMine(bool) }
+
+	local Hash = spellSet.Hash
+	for identifier, mightBeMine in next, updatedAuras do
+		if Hash[identifier] and (mightBeMine or not OnlyMine) then
+			return true
+		end
+	end
+end
+
 function ConditionObject:GenerateUnitAuraString(unit, spell, onlyMine)
+	local unitCheck = type(unit) == "table"
+		-- unit is a UnitSet:
+		and ("%s[arg1]"):format(CNDT:GetTableSubstitution(unit.UnitsLookup))
+		-- unit is a unitID:
+		or ("arg1 == %q"):format(unit)
+
+
 	if TMW.COMMON.Auras then
 		local canUsePacked, auraEvent = TMW.COMMON.Auras:RequestUnits(unit)
 		
 		self:RequestEvent(auraEvent)
 		self:SetNumEventArgs(2)
 
-		local str = ("event == %q and arg1 == %q"):format(auraEvent, unit)
-			-- arg2 is payload:
-			-- If it is nil, the event is a general update for the unit
-			-- If it is a table, its keys are names/ids of what changed
-			-- and the values indicate if the aura might have been mine.
-			.. " and (not arg2 or arg2["
+		if type(spell) == "table" then
+			-- If we were passed a spell set that only contains one spell,
+			-- inline that spell so we can avoid the function call to IsTmwAuraEventRelevant
+			if #spell.Array == 1 then
+				spell = spell.First
+			end
+		end
+
+		local str = ("event == %q and %s and"):format(auraEvent, unitCheck)
+
+		-- arg2 is payload:
+		-- If it is nil, the event is a general update for the unit.
+		-- If it is a table, its keys are names/ids of what changed
+		-- and the values indicate if the aura might have been mine.
+		if type(spell) == "table" then
+			-- Spell is a SpellSet 
+			str = str .. " (not arg2 or ConditionObject:IsTmwAuraEventRelevant(arg2,"
+			.. CNDT:GetTableSubstitution(spell) 
+			.. ", "
+			.. tostring(onlyMine)
+			.. "))"
+		else
+			-- Spell is a name/id
+			str = str .. " (not arg2 or arg2["
 			.. (type(spell) == "string" and format("%q", spell) or tostring(spell))
 			.. "]"
 			.. (onlyMine and "" or " ~= nil")
 			.. ")"
+		end
 		
 		return str
 	else
 		self:RequestEvent("UNIT_AURA")
 		self:SetNumEventArgs(3)
 			
-		local str = "event == 'UNIT_AURA' and arg1 == "
-			.. format("%q", unit)
+		local str = ("event == %q and %s"):format("UNIT_AURA", unitCheck)
+
+		if type(spell) == "string" then
 			-- arg2 is isFullUpdate:
 			-- If it is nil, the client doesn't support the new UNIT_AURA payload.
 			-- If it is true, the event isn't about any particular aura.
 			-- If it is false, then arg3 is `updatedAuras`.
-			.. " and (arg2 ~= false or ConditionObject:IsUnitAuraEventRelevant(arg3,"
+			str = str .. " and (arg2 ~= false or ConditionObject:IsUnitAuraEventRelevant(arg3,"
 			.. (type(spell) == "string" and format("%q", spell) or tostring(spell))
 			.. ", "
 			.. tostring(onlyMine)
 			.. "))"
+		end
 		
 		return str
 	end
@@ -476,7 +515,13 @@ end
 --     ConditionObject:GenerateNormalEventString("UNIT_HEALTH", CNDT:GetUnit(c.Unit))
 -- end,
 function ConditionObject:GetUnitChangedEventString(unit)
-	if unit == "player" then
+	if type(unit) == "table" then
+		-- unit is a UnitSet
+		if not unit.allUnitsChangeOnEvent then 
+			return false 
+		end
+		return unit.event
+	elseif unit == "player" then
 		-- Returning false (as a string, not a boolean) won't cause responses to any events,
 		-- and it also won't make the ConditionObject default to being OnUpdate driven.